Laptop sprint: ops scripts, static site, dashboard attack map, soak evidence.
Consolidate install/verify/soak gates, Cloudflare Pages site assets, security
profiles, deb packaging, and dashboard geo map API — ready for VPS webhook next.

+++ ipc_auth.h
@@ -10,11 +10,11 @@ extern "C" {
 
 void     ipc_auth_init(void);
 void     ipc_auth_load_env_file(const char *path);
-uint32_t ipc_auth_token_for_message(void);
+uint64_t ipc_auth_token_for_message(void);
 int      ipc_auth_required(void);
-int      ipc_auth_validate_message(uint32_t token);
+int      ipc_auth_validate_message(uint64_t token);
 int      ipc_auth_peer_allowed(uid_t uid, gid_t gid);
-int      ipc_auth_shutdown_allowed(uid_t uid, uint32_t token);
+int      ipc_auth_shutdown_allowed(uid_t uid, uint64_t token);
 
 #ifdef __cplusplus
 }
